Refactor Ui and UiApp
Split Responsibility of UiApp with Ui.

UiApp is responsible for launching and styling the GUI as well as for handling user inputs and changing the GUI dynamically according to input.

Ui now not only serves as a message holder but also to help the GUI handle inputs, and getting a response from Butler.

-Shifted handleInputs and static resources from UiApp to Ui

-Fixed a bug where inputting an unformatted date and time would cause the date and time to be added to the String representation of a task object