Add Java Assertions to code in util package
Add assertions to util classes that handle program.

There are no explicit checks in the program to ensure that objects created by manager classes are properly instantiated.

There is no guarantee that during runtime unchecked exceptions such as NullPointerExceptions will not be thrown.

Java assertions were added in util package classes to make sure that objects and object parameters are not null and are properly created by their relevant classes.

This helps to ensure that in the future, bugs and unexpected exceptions during runtime is minimised.