Refactor parse method
Shorten parse method

Parse method is getting very long as there are many keywords for commands to detect.

The long method results in poor code quality as reading the method is difficult, making maintenance difficult as well.

Extracting the different commands that we need to detect allows for better readability for readers as well as better maintainability, as it is easier to debug due to modularity.
This generally improves the quality of the code.